# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Bakery Service API",
...
    "version": "1.0.0"
  },
  "paths": {
    "/v1/customers": {
      "get": {
        "tags": [
          "Customers"
        ],
        "summary": "Daftar Seluruh Pelanggan",
        "description": "Mengambil seluruh data profil pelanggan yang terdaftar pada sistem toko roti.",
        "operationId": "list_customers_v1_customers_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerListResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Customers"
        ],
        "summary": "Mendaftarkan Pelanggan Baru",
        "description": "Membuat data profil pelanggan baru ke dalam sistem.",
        "operationId": "create_customer_v1_customers_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomerCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerResponse"
                }
              }
            }
          },
          "400": {
            "description": "Parameter permintaan tidak valid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RFCErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/products": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "Daftar Produk",
        "description": "Mengambil daftar seluruh produk beserta informasi harga dan stok yang tersedia. Dapat difilter berdasarkan kategori produk.",
        "operationId": "list_products_v1_products_get",
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter berdasarkan kategori produk (misal: 'bread', 'pastry')",
              "examples": [
                "bread"
              ],
              "title": "Category"
            },
            "description": "Filter berdasarkan kategori produk (misal: 'bread', 'pastry')"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/products/{product_id}/stock": {
      "put": {
        "tags": [
          "Products"
        ],
        "summary": "Perbarui Stok Produk [Idempoten - HTTP PUT]",
        "description": "Menetapkan nilai stok absolut untuk suatu produk roti. Operasi HTTP PUT ini bersifat **Idempoten** karena jika dipanggil berulang kali dengan nilai stok yang sama, nilai akhir stok pada sistem tetap sama.",
        "operationId": "update_product_stock_v1_products__product_id__stock_put",
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "ID unik produk",
              "examples": [
                "prod_sourdough01"
              ],
              "title": "Product Id"
            },
            "description": "ID unik produk"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductStockUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponse"
                }
              }
            }
          },
          "404": {
            "description": "Produk tidak ditemukan",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RFCErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/orders": {
      "post": {
        "tags": [
          "Orders"
        ],
        "summary": "Membuat Pesanan [Idempoten - Pola Stripe]",
        "description": "**[Operasi Idempoten - Pola Stripe]**\n\nMembuat transaksi pesanan baru dan melakukan alokasi pengurangan stok produk secara otomatis.\n\nEndpoint ini mewajibkan/mendukung header `Idempotency-Key` (UUID v4) guna menjamin tidak terjadinya duplikasi order dan pengurangan stok berganda apabila terjadi pengulangan permintaan dari klien. Pengulangan permintaan dengan kunci sama dan payload sama akan mengembalikan respons *cache* (dengan header `Idempotent-Replay: true` dan `X-Cache-Lookup: HIT`). Jika kunci yang sama dipakai dengan payload berbeda, server mengembalikan status `HTTP 409 Conflict`.",
        "operationId": "create_order_v1_orders_post",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Kunci unik (UUID v4) untuk mencegah duplikasi pemrosesan pesanan",
              "title": "Idempotency-Key"
            },
            "description": "Kunci unik (UUID v4) untuk mencegah duplikasi pemrosesan pesanan"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Pesanan berhasil dibuat",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponse"
                }
              }
            }
          },
          "400": {
            "description": "Parameter tidak valid atau stok tidak mencukupi",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RFCErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Pelanggan atau Produk tidak ditemukan",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RFCErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Konflik Idempotency-Key digunakan dengan parameter berbeda",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RFCErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Orders"
        ],
        "summary": "Daftar Pesanan",
        "description": "Mengambil daftar transaksi pesanan dengan dukungan navigasi kursor (cursor-based pagination) dan filter status.",
        "operationId": "list_orders_v1_orders_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Jumlah data per halaman (1–100)",
              "examples": [
                10
              ],
              "default": 10,
              "title": "Limit"
            },
            "description": "Jumlah data per halaman (1–100)"
          },
          {
            "name": "starting_after",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "ID pesanan sebagai kursor untuk mengambil data setelah entri ini",
              "examples": [
                "ord_20260925_001"
              ],
              "title": "Starting After"
            },
            "description": "ID pesanan sebagai kursor untuk mengambil data setelah entri ini"
          },
          {
            "name": "ending_before",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "ID pesanan sebagai kursor untuk mengambil data sebelum entri ini",
              "title": "Ending Before"
            },
            "description": "ID pesanan sebagai kursor untuk mengambil data sebelum entri ini"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter berdasarkan status pesanan (misal: pending_payment, paid, completed)",
              "title": "Status"
            },
            "description": "Filter berdasarkan status pesanan (misal: pending_payment, paid, completed)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/orders/{order_id}/cancel": {
      "post": {
        "tags": [
          "Orders"
        ],
        "summary": "Membatalkan Pesanan [Idempoten - State Transition]",
        "description": "**[Operasi Idempoten - State Transition & Idempotency-Key]**\n\nMembatalkan pesanan yang belum selesai serta mengembalikan alokasi kuota stok produk.\nMendukung header `Idempotency-Key` dan transisi status aman sehingga pemanggilan berulang tidak akan mengembalikan stok roti lebih dari 1 kali.",
        "operationId": "cancel_order_v1_orders__order_id__cancel_post",
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "ID unik pesanan yang akan dibatalkan",
              "examples": [
                "ord_20260925_002"
              ],
              "title": "Order Id"
            },
            "description": "ID unik pesanan yang akan dibatalkan"
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Kunci unik UUID v4",
              "title": "Idempotency-Key"
            },
            "description": "Kunci unik UUID v4"
          }
        ],
        "responses": {
          "200": {
            "description": "Pesanan berhasil dibatalkan",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponse"
                }
              }
            }
          },
          "400": {
            "description": "Status pesanan tidak memenuhi syarat untuk dibatalkan",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RFCErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Pesanan tidak ditemukan",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RFCErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Konflik kunci idempotensi",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RFCErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/payments": {
      "post": {
        "tags": [
          "Payments"
        ],
        "summary": "Memproses Pembayaran [Idempoten - Pola Stripe]",
        "description": "**[Operasi Idempoten - Pola Stripe Finansial]**\n\nMemproses transaksi pembayaran pesanan toko roti. Operasi ini dilindungi oleh header `Idempotency-Key` (UUID v4) guna menjamin bahwa pelanggan tidak akan terdebit dua kali (*double charging*) saat terjadi kegagalan/timeout jaringan. Pengulangan permintaan dengan kunci sama akan mengembalikan respons *cache* (dengan header `Idempotent-Replay: true` dan `X-Cache-Lookup: HIT`).",
        "operationId": "create_payment_v1_payments_post",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Kunci unik UUID v4 untuk mencegah transaksi pembayaran ganda",
              "title": "Idempotency-Key"
            },
            "description": "Kunci unik UUID v4 untuk mencegah transaksi pembayaran ganda"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaymentCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Pembayaran berhasil diproses",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentResponse"
                }
              }
            }
          },
          "400": {
            "description": "Nominal pembayaran tidak sesuai atau status pesanan tidak valid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RFCErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Pesanan tidak ditemukan",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RFCErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Konflik kunci idempotensi",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RFCErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/payments/{payment_id}/confirm": {
      "post": {
        "tags": [
          "Payments"
        ],
        "summary": "Konfirmasi Pembayaran [Idempoten - State Transition]",
        "description": "**[Operasi Idempoten - State Transition]**\n\nMelakukan konfirmasi status penyelesaian transaksi pembayaran. Operasi ini idempoten karena menetapkan status akhir ke `settlement`. Pemanggilan berulang tidak mengubah hasil akhir.",
        "operationId": "confirm_payment_v1_payments__payment_id__confirm_post",
        "parameters": [
          {
            "name": "payment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "ID unik pembayaran",
              "examples": [
                "pay_trx_001"
              ],
              "title": "Payment Id"
            },
            "description": "ID unik pembayaran"
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Kunci unik UUID v4",
              "title": "Idempotency-Key"
            },
            "description": "Kunci unik UUID v4"
          }
        ],
        "responses": {
          "200": {
            "description": "Pembayaran dikonfirmasi",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentResponse"
                }
              }
            }
          },
          "404": {
            "description": "Pembayaran tidak ditemukan",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RFCErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CustomerCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "examples": [
              "Budi Santoso"
            ]
          },
          "phone": {
            "type": "string",
            "title": "Phone",
            "examples": [
              "081234567890"
            ]
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email",
            "examples": [
              "budi@example.com"
            ]
          }
        },
        "type": "object",
        "required": [
          "name",
          "phone"
        ],
        "title": "CustomerCreate"
      },
      "CustomerListResponse": {
        "properties": {
          "object": {
            "type": "string",
            "title": "Object",
            "default": "list",
            "examples": [
              "list"
            ]
          },
          "url": {
            "type": "string",
            "title": "Url",
            "default": "/v1/customers",
            "examples": [
              "/v1/customers"
            ]
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More",
            "examples": [
              false
            ]
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/CustomerResponse"
            },
            "type": "array",
            "title": "Data"
          }
        },
        "type": "object",
        "required": [
          "has_more",
          "data"
        ],
        "title": "CustomerListResponse"
      },
      "CustomerResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "examples": [
              "cus_9a8b1c"
            ]
          },
          "object": {
            "type": "string",
            "title": "Object",
            "default": "customer",
            "examples": [
              "customer"
            ]
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "phone": {
            "type": "string",
            "title": "Phone"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "created": {
            "type": "integer",
            "title": "Created"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "phone",
          "email"
        ],
        "title": "CustomerResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "OrderCreate": {
        "properties": {
          "customer_id": {
            "type": "string",
            "title": "Customer Id",
            "examples": [
              "cus_budi01"
            ]
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/OrderItemInput"
            },
            "type": "array",
            "title": "Items"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes",
            "examples": [
              "Potong slice tebal"
            ]
          }
        },
        "type": "object",
        "required": [
          "customer_id",
          "items"
        ],
        "title": "OrderCreate"
      },
...
      "OrderResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "examples": [
              "ord_20260925_001"
            ]
          },
          "object": {
            "type": "string",
            "title": "Object",
            "default": "order",
            "examples": [
              "order"
            ]
          },
          "customer": {
            "type": "string",
            "title": "Customer",
            "examples": [
              "cus_budi01"
            ]
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "Status pesanan: pending_payment, paid, baking, ready_for_pickup, completed, cancelled.",
            "examples": [
              "pending_payment"
            ]
          },
          "currency": {
            "type": "string",
            "title": "Currency",
            "default": "idr",
            "examples": [
              "idr"
            ]
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/OrderItemDetail"
            },
            "type": "array",
            "title": "Items"
          },
          "total_amount": {
            "type": "integer",
            "title": "Total Amount",
            "examples": [
              90000
            ]
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "created": {
            "type": "integer",
            "title": "Created"
          }
        },
        "type": "object",
        "required": [
          "id",
          "customer",
          "status",
          "items",
          "total_amount"
        ],
        "title": "OrderResponse"
      },
...
      "ProductListResponse": {
        "properties": {
          "object": {
            "type": "string",
            "title": "Object",
            "default": "list",
            "examples": [
              "list"
            ]
          },
          "url": {
            "type": "string",
            "title": "Url",
            "default": "/v1/products",
            "examples": [
              "/v1/products"
            ]
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More",
            "examples": [
              false
            ]
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/ProductResponse"
            },
            "type": "array",
            "title": "Data"
          }
        },
        "type": "object",
        "required": [
          "has_more",
          "data"
        ],
        "title": "ProductListResponse"
      },
      "ProductResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "examples": [
              "prod_sourdough01"
            ]
          },
          "object": {
            "type": "string",
            "title": "Object",
            "default": "product",
            "examples": [
              "product"
            ]
          },
          "name": {
            "type": "string",
            "title": "Name",
            "examples": [
              "Artisan Sourdough Country Loaf"
            ]
          },
          "category": {
            "type": "string",
            "title": "Category",
            "examples": [
              "bread"
            ]
          },
          "unit_price": {
            "type": "integer",
            "title": "Unit Price",
            "description": "Harga dalam satuan IDR.",
            "examples": [
              45000
            ]
          },
          "stock": {
            "type": "integer",
            "title": "Stock",
            "description": "Jumlah ketersediaan stok.",
            "examples": [
              20
            ]
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "examples": [
              "Roti sourdough fermentasi 24 jam alami tanpa ragi instan."
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "category",
          "unit_price",
          "stock"
        ],
        "title": "ProductResponse"
      },
      "ProductStockUpdate": {
        "properties": {
          "stock": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Stock",
            "description": "Jumlah stok absolut yang ditetapkan",
            "examples": [
              50
            ]
          }
        },
        "type": "object",
        "required": [
          "stock"
        ],
        "title": "ProductStockUpdate"
      },
      "RFCErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "RFCErrorResponse"
      },
...
    }
  }
}